Refactor scan_url method
Each API offers the fetch_subdomains method in
the latest version of ReconLib. All we need to
do now is call the method in a collection of
API instances. I've made the method return a set
instead of a list (like before) so we can have
automatic elimination of duplicate results,
which is a common occurrence when dealing with
multiple services.